+++ src/airalogy/built_in_types/file.py
@@ -54,3 +54,5 @@ FileIdXLSX = Annotated[str, FileIdField("xlsx")]
 FileIdPPTX = Annotated[str, FileIdField("pptx")]
 
 FileIdPDF = Annotated[str, FileIdField("pdf")]
+
+FileIdDNA = Annotated[str, FileIdField("dna")]